Improve security by **requiring** passwords
The change improves the security of the application by having the
default password be set to an invalid value of `null` instead of a valid
value of `"password"`. The change uses a special template
function---`required`---to prevent the installation of the chart without
the user providing a valid password.