Allow `root` to connect to MySQL from any host
The change is introducing a potential security weakness by allowing the
`root` user to connect from any host. However, the `mysql` service is
only accessible from nodes in the Kubernetes cluster and the cluster is
inaccessible to the outside world. The appropriate configuration would
be to address the cluster IP address as the value of `MYSQL_ROOT_HOST`.